backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
PyGithub==2.6.1
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.11.7
